Moving startuplogo to be clearer it's not cli specific; moving rootcommand to Program.cs to make bootstrapping clearer
